@@ -4,9 +4,12 @@ package main
 // see comments in lockd.go
 //
 
-import "6.5840/lockservice"
-import "os"
-import "fmt"
+import (
+	"fmt"
+	"os"
+
+	"6.5840/lockservice"
+)
 
 func usage() {
 	fmt.Printf("Usage: lockc -l|-u primaryport backupport lockname\n")
